proxyAPI/requirements.txt
fastapi
pymongo
langchain
langchain-mongodb
langchain-openai
langchain-community
python-dotenv
httpx
pydantic
uvicorn
